# Frontend/RepoGeist/openapi.json
{
    "openapi":  "3.1.1",
    "info":  {
                 "title":  "RepoGeist | v1",
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:5221/"
                    }
                ],
    "paths":  {
                  "/api/repos/{owner}/{name}/analyze":  {
                                                            "post":  {
                                                                         "tags":  [
                                                                                      "Analysis"
                                                                                  ],
                                                                         "parameters":  [
                                                                                            {
                                                                                                "name":  "owner",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  {
                                                                                                               "type":  "string"
                                                                                                           }
                                                                                            },
                                                                                            {
                                                                                                "name":  "name",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  {
                                                                                                               "type":  "string"
                                                                                                           }
                                                                                            }
                                                                                        ],
                                                                         "responses":  {
                                                                                           "200":  {
                                                                                                       "description":  "OK",
                                                                                                       "content":  {
                                                                                                                       "text/plain":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/AnalysisStatusDto"
                                                                                                                                                     }
                                                                                                                                      },
                                                                                                                       "application/json":  {
                                                                                                                                                "schema":  {
                                                                                                                                                               "$ref":  "#/components/schemas/AnalysisStatusDto"
                                                                                                                                                           }
                                                                                                                                            },
                                                                                                                       "text/json":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "$ref":  "#/components/schemas/AnalysisStatusDto"
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                                   }
                                                                                                   }
                                                                                       }
                                                                     }
                                                        },
                  "/api/repos/{owner}/{name}/analysis-status":  {
                                                                    "get":  {
                                                                                "tags":  [
                                                                                             "Analysis"
                                                                                         ],
                                                                                "parameters":  [
                                                                                                   {
                                                                                                       "name":  "owner",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "string"
                                                                                                                  }
                                                                                                   },
                                                                                                   {
                                                                                                       "name":  "name",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "string"
                                                                                                                  }
                                                                                                   }
                                                                                               ],
                                                                                "responses":  {
                                                                                                  "200":  {
                                                                                                              "description":  "OK",
                                                                                                              "content":  {
                                                                                                                              "text/plain":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/AnalysisStatusDto"
                                                                                                                                                            }
                                                                                                                                             },
                                                                                                                              "application/json":  {
                                                                                                                                                       "schema":  {
                                                                                                                                                                      "$ref":  "#/components/schemas/AnalysisStatusDto"
                                                                                                                                                                  }
                                                                                                                                                   },
                                                                                                                              "text/json":  {
                                                                                                                                                "schema":  {
                                                                                                                                                               "$ref":  "#/components/schemas/AnalysisStatusDto"
                                                                                                                                                           }
                                                                                                                                            }
                                                                                                                          }
                                                                                                          },
                                                                                                  "404":  {
                                                                                                              "description":  "Not Found",
                                                                                                              "content":  {
                                                                                                                              "text/plain":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                            }
                                                                                                                                             },
                                                                                                                              "application/json":  {
                                                                                                                                                       "schema":  {
                                                                                                                                                                      "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                                  }
                                                                                                                                                   },
                                                                                                                              "text/json":  {
                                                                                                                                                "schema":  {
                                                                                                                                                               "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                           }
                                                                                                                                            }
                                                                                                                          }
                                                                                                          }
                                                                                              }
                                                                            }
                                                                },
                  "/api/repos/{owner}/{name}/graph":  {
                                                          "get":  {
                                                                      "tags":  [
                                                                                   "Analysis"
                                                                               ],
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "owner",
                                                                                             "in":  "path",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                         },
                                                                                         {
                                                                                             "name":  "name",
                                                                                             "in":  "path",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "OK",
                                                                                                    "content":  {
                                                                                                                    "text/plain":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/DependencyGraphDto"
                                                                                                                                                  }
                                                                                                                                   },
                                                                                                                    "application/json":  {
                                                                                                                                             "schema":  {
                                                                                                                                                            "$ref":  "#/components/schemas/DependencyGraphDto"
                                                                                                                                                        }
                                                                                                                                         },
                                                                                                                    "text/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/DependencyGraphDto"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                                }
                                                                                                },
                                                                                        "404":  {
                                                                                                    "description":  "Not Found",
                                                                                                    "content":  {
                                                                                                                    "text/plain":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                  }
                                                                                                                                   },
                                                                                                                    "application/json":  {
                                                                                                                                             "schema":  {
                                                                                                                                                            "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                        }
                                                                                                                                         },
                                                                                                                    "text/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                                }
                                                                                                }
                                                                                    }
                                                                  }
                                                      },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "AnalysisStatusDto":  {
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "repoId":  {
                                                                                                   "type":  "string",
                                                                                                   "format":  "uuid"
                                                                                               },
                                                                                    "jobId":  {
                                                                                                  "type":  "string",
                                                                                                  "format":  "uuid"
                                                                                              },
                                                                                    "status":  {
                                                                                                   "type":  "string"
                                                                                               },
                                                                                    "progressPercent":  {
                                                                                                            "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                                            "type":  [
                                                                                                                         "null",
                                                                                                                         "integer",
                                                                                                                         "string"
                                                                                                                     ],
                                                                                                            "format":  "int32"
                                                                                                        },
                                                                                    "errorMessage":  {
                                                                                                         "type":  [
                                                                                                                      "null",
                                                                                                                      "string"
                                                                                                                  ]
                                                                                                     }
                                                                                }
                                                             },
...
                                       "DependencyGraphDto":  {
                                                                  "type":  "object",
                                                                  "properties":  {
                                                                                     "repoId":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "uuid"
                                                                                                },
                                                                                     "nodes":  {
                                                                                                   "type":  "array",
                                                                                                   "items":  {
                                                                                                                 "$ref":  "#/components/schemas/GraphNodeDto"
                                                                                                             }
                                                                                               },
                                                                                     "edges":  {
                                                                                                   "type":  "array",
                                                                                                   "items":  {
                                                                                                                 "$ref":  "#/components/schemas/GraphEdgeDto"
                                                                                                             }
                                                                                               }
                                                                                 }
                                                              },
...
                                       "GraphNodeDto":  {
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "id":  {
                                                                                          "type":  "string",
                                                                                          "format":  "uuid"
                                                                                      },
                                                                               "nodeType":  {
                                                                                                "type":  "string"
                                                                                            },
                                                                               "pathOrName":  {
                                                                                                  "type":  "string"
                                                                                              }
                                                                           }
                                                        },
...
                                   }
                   },
...
}